theater_cli::error

Enum CliError

pub enum CliError {

    ConnectionFailed {
        address: SocketAddr,
        source: Error,
    },
    ConnectionLost,
    ConnectionTimeout {
        timeout: u64,
    },
    ActorNotFound {
        actor_id: String,
    },
    ActorStartFailed {
        actor_id: String,
        reason: String,
    },
    ActorNotRunning {
        actor_id: String,
    },
    ActorError {
        actor_id: String,
        reason: String,
    },
    ActorStopped {
        actor_id: String,
        reason: String,
    },
    InvalidProjectDirectory {
        path: String,
    },
    BuildFailed {
        output: String,
    },
    MissingTool {
        tool: String,
        install_command: String,
    },
    InvalidManifest {
        reason: String,
    },
    ConfigError {
        reason: String,
    },
    TemplateNotFound {
        template: String,
        available: String,
    },
    TemplateError {
        reason: String,
    },
    FileOperationFailed {
        operation: String,
        path: String,
        source: Error,
    },
    PermissionDenied {
        operation: String,
        path: String,
    },
    InvalidInput {
        field: String,
        value: String,
        suggestion: String,
    },
    ValidationError {
        reason: String,
    },
    ServerError {
        message: String,
    },
    ProtocolError {
        reason: String,
    },
    UnexpectedResponse {
        response: String,
    },
    EventStreamError {
        reason: String,
    },
    EventFilterError {
        filter: String,
    },
    OperationCancelled,
    Internal(Error),
    Serialization(Error),
    Io(Error),
    NetworkError {
        operation: String,
        source: Box<dyn Error + Send + Sync>,
    },
    InvalidResponse {
        message: String,
        source: Option<Box<dyn Error + Send + Sync>>,
    },
    IoError {
        operation: String,
        source: Error,
    },
    ParseError {
        message: String,
    },
    NotImplemented {
        feature: String,
        message: String,
    },
    ManagementError(ManagementError),

}

Main error type for the Theater CLI

Variants

ConnectionFailed

Connection-related errors

Fields

address: SocketAddr

source: Error

ConnectionLost

ConnectionTimeout

Fields

timeout: u64

ActorNotFound

Actor-related errors

Fields

actor_id: String

ActorStartFailed

Fields

actor_id: String

reason: String

ActorNotRunning

Fields

actor_id: String

ActorError

Fields

actor_id: String

reason: String

ActorStopped

Fields

actor_id: String

reason: String

InvalidProjectDirectory

Project and build errors

Fields

path: String

BuildFailed

Fields

output: String

MissingTool

Fields

tool: String

install_command: String

InvalidManifest

Manifest and configuration errors

Fields

reason: String

ConfigError

Fields

reason: String

TemplateNotFound

Template errors

Fields

template: String

available: String

TemplateError

Fields

reason: String

FileOperationFailed

I/O and filesystem errors

Fields

operation: String

path: String

source: Error

PermissionDenied

Fields

operation: String

path: String

InvalidInput

Validation errors

Fields

field: String

value: String

suggestion: String

ValidationError

Fields

reason: String

ServerError

Server and protocol errors

Fields

message: String

ProtocolError

Fields

reason: String

UnexpectedResponse

Fields

response: String

EventStreamError

Event and monitoring errors

Fields

reason: String

EventFilterError

Fields

filter: String

OperationCancelled

Cancellation errors

Internal(Error)

Generic wrapper for other errors

Serialization(Error)

Io(Error)

NetworkError

Network and protocol specific errors

Fields

operation: String

source: Box<dyn Error + Send + Sync>

InvalidResponse

Fields

message: String

source: Option<Box<dyn Error + Send + Sync>>

IoError

Fields

operation: String

source: Error

ParseError

Fields

message: String

NotImplemented

Fields

feature: String

message: String

ManagementError(ManagementError)

Implementations

impl CliError

pub fn connection_failed(address: SocketAddr, source: impl Into<Error>) -> Self

Create a connection failed error

pub fn actor_not_found(actor_id: impl Into<String>) -> Self

Create an actor not found error

pub fn build_failed(output: impl Into<String>) -> Self

Create a build failed error

pub fn invalid_manifest(reason: impl Into<String>) -> Self

Create an invalid manifest error

pub fn template_not_found( template: impl Into<String>, available: Vec<String>, ) -> Self

Create a template not found error

pub fn file_operation_failed( operation: impl Into<String>, path: impl Into<String>, source: Error, ) -> Self

Create a file operation failed error

pub fn invalid_input( field: impl Into<String>, value: impl Into<String>, suggestion: impl Into<String>, ) -> Self

Create an invalid input error with helpful suggestions

pub fn invalid_actor_id(actor_id: impl Into<String>) -> Self

Create an invalid actor ID error

pub fn operation_timeout(_operation: impl Into<String>, timeout: u64) -> Self

pub fn not_implemented( feature: impl Into<String>, message: impl Into<String>, ) -> Self

Create a not implemented error

pub fn management_error(error: ManagementError) -> Self

Create a management error from a Theater server error

pub fn user_message(&self) -> String

Get a user-friendly error message with potential solutions

pub fn is_retryable(&self) -> bool

Check if this error suggests the user should retry

pub fn category(&self) -> &'static str

Get the error category for metrics/logging